+++ src/main/java/com/contractor/model/enums/UserRights.java
@@ -1,0 +1,17 @@
+package com.contractor.model.enums;
+
+/**
+ * Created by dpecanac on 29.8.2017.
+ */
+public enum UserRights {
+    deleteJob,
+    deleteUser,
+    getJob,
+    getUser,
+    login,
+    logout,
+    postJob,
+    postUser,
+    putJob,
+    putUser,
+}